Updated versioning
The listings and combos and feature associations
are much tougher than anticipated. Creation of a
separate listing index for the milestone manage window
is necessary.
Also simplified the version numbering by truncating the
patch level. That does not make sense in a minimal
viable product.
Also, removed the version from the project creation/edit
widget. Makes absolutely no sense, as it's managed by
the roadmap widget and would only cause confusion.
E.g. edit current version.. Whatever would that do? Milestones
are interconnected with the features and issues, so the milestone
manager should only be capable of changing them.