research: replace milestone push button
This is untested. But current assumption is
that we can use something like this to replace
the stock push button in the roadmap.ui for milestone
selection with an MPushButton instance generated
from code.

This is necessary, because we need to intercept the
sender() in the push button class, which is difficult
with the imported ui.