Re-factoring data structure NOT EXECUTING
This is a progress commit for a bigger internal structure
overhaul. It's not executing, as much of the code is not
added yet.

It features a new class: DataStore, that replaces savdat
and rundat. This will take care of saving / loading.

Also remove the global rundat index (most of it) and use
local (self) addressing. It's ugly and I was not very
consequent with it anyway.

Global overhaul will help with roadmap complexity (by reducing
systemic complexity and uniting data addressing).