config frame, save on exit and load last
The new configuration system is now initialized at startup
(mainwindow). It creates the ~/.config/nelia1 directory if
not yet present. Then goes on to load the last configuration
if existent. If it finds a valid path, it will also enable
the "Open Last" button in the project widget, which loads
the last project and then is disabled. The lastpath variable
is set on save and load.

The configuration and the project file (if path existent) are
saved automatically on exit. If not saved yet, it will prompt
the user to specify a filename to save. (verified?)