added fail dialogs for save and load
For save, it now shows a fail dialog if save failed. This most
likely happens when file access is restricted. But the implementation
is catch all exceptions. It will reset the path and return False, which
might or might not be handled. Same goes for load method. This also
catches file parsing errors (i.e. invalid file contents).