merged milestone into roadmap
as the milestone has been moved to the new datacore data model,
and the roadmap assumes that the old ones are still in place, this
version (and the previous ones since the last 0.9 are not handling
milestone items. To fix this, the roadmap (and the mpushbutton have to
be migrated.