selected project label and project in table
There is a bug with the Ubuntu theme that prevents from displaying selected rows
in the table in a distinctive way when not focused. So I decided to add a
selected project label to the top that will update in real time.

Also added preliminary line listing in the project list. I need to change this,
because it does not respect the sorting yet.