worked around focus loss issue
So when using keyboard only, the table index focus works. When using the mouse
to select items in the table QSelectionModel.currentIndex() starts returning -1
even though a table row is most defilately selected (shown visually in the gui).
This was a particularily obnoxious bug during most of the time.

This strikes out the last of the dev todo list. Beta 3 it is.